+++ changelog.md
@@ -1,7 +1,9 @@
 ### Revision History
 * 1.1.4
-  * Added `server.tomcat.max-threads=200` to default `application.properties`
   * Consumed n-cube 4.1.5
+  * Added `server.tomcat.max-threads=200` to default `application.properties`
+  * Modal updates for search options and delete/restore cubes; code cleanup.
+  * Bug fix: form-builder not properly adding table rows from add button.
 * 1.1.3
   * Consumed n-cube 4.1.4
 * 1.1.2